+++ crates/jxl-render/src/features/mod.rs
@@ -1,7 +1,9 @@
 mod noise;
 mod spline;
+mod spot_colors;
 mod upsampling;
 
 pub use noise::render_noise;
 pub use spline::render_spline;
+pub use spot_colors::render_spot_color;
 pub use upsampling::upsample;
