+++ crates/jxl-render/src/state.rs
@@ -1,0 +1,79 @@
+use std::collections::HashMap;
+
+use jxl_frame::data::{LfGlobal, HfGlobal, LfGroup};
+use jxl_modular::ChannelShift;
+
+use crate::region::ImageWithRegion;
+
+#[derive(Debug)]
+pub struct RenderState {
+    pub(crate) renders: Vec<FrameRender>,
+    pub(crate) loading_render_cache: Option<RenderCache>,
+}
+
+impl RenderState {
+    pub fn new() -> Self {
+        Self {
+            renders: Vec::new(),
+            loading_render_cache: None,
+        }
+    }
+}
+
+impl RenderState {
+    pub fn preserve_current_frame(&mut self) {
+        if let Some(cache) = self.loading_render_cache.take() {
+            self.renders.push(FrameRender::InProgress(Box::new(cache)));
+        } else {
+            self.renders.push(FrameRender::None);
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct RenderCache {
+    pub(crate) lf_global: Option<LfGlobal>,
+    pub(crate) hf_global: Option<HfGlobal>,
+    pub(crate) lf_groups: HashMap<u32, LfGroup>,
+}
+
+impl RenderCache {
+    pub fn new(frame: &crate::IndexedFrame) -> Self {
+        let frame_header = frame.header();
+        let jpeg_upsampling = frame_header.jpeg_upsampling;
+        let shifts_cbycr: [_; 3] = std::array::from_fn(|idx| {
+            ChannelShift::from_jpeg_upsampling(jpeg_upsampling, idx)
+        });
+
+        let lf_width = (frame_header.color_sample_width() + 7) / 8;
+        let lf_height = (frame_header.color_sample_height() + 7) / 8;
+        let mut whd = [(lf_width, lf_height); 3];
+        for ((w, h), shift) in whd.iter_mut().zip(shifts_cbycr) {
+            let (shift_w, shift_h) = shift.shift_size((lf_width, lf_height));
+            *w = shift_w;
+            *h = shift_h;
+        }
+        Self {
+            lf_global: None,
+            hf_global: None,
+            lf_groups: HashMap::new(),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum FrameRender {
+    None,
+    InProgress(Box<RenderCache>),
+    Done(ImageWithRegion),
+}
+
+impl FrameRender {
+    pub fn as_grid(&self) -> Option<&ImageWithRegion> {
+        if let Self::Done(grid) = self {
+            Some(grid)
+        } else {
+            None
+        }
+    }
+}
